+++ pkg/visor/rpc.go
@@ -339,6 +339,7 @@ func (r *RPC) SetAppPK(in *SetAppPKIn, _ *struct{}) (err error) {
 	return r.visor.setAppPK(in.AppName, in.PK)
 }
 
+// GetAppConnectionsSummary returns connections stats for the app.
 func (r *RPC) GetAppConnectionsSummary(appName *string, out *[]appserver.ConnectionSummary) (err error) {
 	defer rpcutil.LogCall(r.log, "GetAppConnectionsSummary", appName)(out, &err)
